[GoalTimeline] Standardize folder structure with index
